factored grammar functions
